Create main+api.py
I aimed to display a CIRCOS plot that has 1) the number of sectors equal to the number of walnut chromosomes, 2) the size of sectors in relative to the size of each walnut chromosome size, and 3) the point plot location for a specific gene (according to the user's geneID input) in relative to the beginning of its genomic range within its chromosome sector. Although I was able to retrieve gene information from NCBI database based on each gene ID, I could not reflect each gene location on CIRCOS plot chromosome sectors.