Add files via upload
This Python code provides the following new functions:
A user can select multiple gene IDs simultaneously.
A plot shows multiple dots that reflect selected gene IDs.
The x-axis of each dot reflects the beginning of a selected gene.
The y-axis of each dot reflects the orientation of a selected gene (when orientation is plus, y = 1; when orientation is minus, y = 0).
Meanwhile, it does not provide the following function:
The full API response displayed on software is well organized according to Dr. Paulo's request.